Moved tests to match new class structure, fixing a few minor niggles along the way.
