+++ generator/src/test/java/com/graphicsfuzz/generator/transformation/AddWrappingConditionalTransformationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.graphicsfuzz.generator.transformation.controlflow;
+package com.graphicsfuzz.generator.transformation;
 
 import com.graphicsfuzz.common.ast.TranslationUnit;
 import com.graphicsfuzz.common.ast.decl.FunctionDefinition;
@@ -34,7 +34,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-public class AddWrappingConditionalStmtsTest {
+public class AddWrappingConditionalTransformationTest {
 
   @Test
   public void testIfProblems() throws Exception {
