+++ generator/src/test/java/com/graphicsfuzz/generator/transformation/SplitForLoopTransformationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.graphicsfuzz.generator.transformation.controlflow;
+package com.graphicsfuzz.generator.transformation;
 
 import com.graphicsfuzz.common.ast.TranslationUnit;
 import com.graphicsfuzz.common.ast.decl.FunctionDefinition;
@@ -60,7 +60,7 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 
-public class SplitForLoopsTest {
+public class SplitForLoopTransformationTest {
 
   @Test
   public void suitableForSplittingNoNext()
